LayoutTests/imported/w3c/web-platform-tests/navigation-api/per-entry-events/dispose-for-navigation-in-child-expected.txt



FAIL Dispose events should fire when entries are removed by a navigation in a different frame assert_equals: expected 0 but got 3
